switch remaining time types to unsigned long (result of millis())

+++ PreheatingAnswer.hpp
@@ -7,23 +7,23 @@
 
 class PreheatingAnswer {
   private:
-    std::vector<long> greenFlashes;
-    std::vector<long> redFlashes;
-    unsigned int CountProcessorFlashesWithLength(std::vector<long> &flashes, unsigned int lengthInMilliseconds);
+    std::vector<unsigned long> greenFlashes;
+    std::vector<unsigned long> redFlashes;
+    unsigned int CountProcessorFlashesWithLength(std::vector<unsigned long> &flashes, unsigned long lengthInMilliseconds);
   public:
-    PreheatingAnswer(std::vector<long> greenFlashes, std::vector<long> redFlashes);
+    PreheatingAnswer(std::vector<unsigned long> greenFlashes, std::vector<unsigned long> redFlashes);
     unsigned int CountRedFlashesWithLength(unsigned long lengthInMilliseconds);
     unsigned int CountGreenFlashesWithLength(unsigned long lengthInMilliseconds);
 };
 
-PreheatingAnswer::PreheatingAnswer(std::vector<long> greenFlashes, std::vector<long> redFlashes) : greenFlashes(greenFlashes), redFlashes(redFlashes) { }
+PreheatingAnswer::PreheatingAnswer(std::vector<unsigned long> greenFlashes, std::vector<unsigned long> redFlashes) : greenFlashes(greenFlashes), redFlashes(redFlashes) { }
 
-unsigned int PreheatingAnswer::CountProcessorFlashesWithLength(std::vector<long> &flashes, unsigned int lengthInMilliseconds) {
-  unsigned int thresh = lengthInMilliseconds * .9;
-  unsigned int minTime = lengthInMilliseconds - thresh;
-  unsigned int maxTime = lengthInMilliseconds + thresh;
+unsigned int PreheatingAnswer::CountProcessorFlashesWithLength(std::vector<unsigned long> &flashes, unsigned long lengthInMilliseconds) {
+  unsigned long thresh = lengthInMilliseconds * .9;
+  unsigned long minTime = lengthInMilliseconds - thresh;
+  unsigned long maxTime = lengthInMilliseconds + thresh;
   unsigned int num = 0;
-  for (long &t : flashes) {
+  for (unsigned long &t : flashes) {
     if (t >= minTime && t <= maxTime) {
       ++num;
     }
